Added pop animation to squares
As a quick fix, since the flashing seems to occur during DOM
manipulation while all new squares are being added, giving the squares
a pop-in animation as an entrance makes it less uncomfortable to the
eyes when scrolling down.
If the pop-in animation is too much, a regular fade in css3 keyframe
animation can easily be made as well.