Convert _ to space in email template
Addresses an issue where configurations with space weren’t accepted.
When setting up a unit file (for use with systemctl on fedora) adding a
space broke parsing of the environment variable. Escaping the space
preserved parsing, but broke JSON. This allows the user to specify
space without actually using space.