Resolve issue with search
In the recent search upgrade, search was switched to be based on the
petitions function, which is used for the front page and doesn’t
display anything that has been updated or responded.

This fix resolves that by creating a new publication and changing the
underlying function to allow ignoring the status of the petition.